requirements.txt
google-generativeai
python-dotenv
openai
gunicorn
flask